Major refactor

+++ TVShows/src/main/java/com/TVShows/DTO/WatchingStatusRequest.java
@@ -6,9 +6,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Data @Builder @AllArgsConstructor @NoArgsConstructor
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
 public class WatchingStatusRequest {
-
-	private Long showId;
+	private Integer showId;
 	private ViewerStatus status;
 }
