+++ TVShows/src/main/java/com/TVShows/repo/NewsArticleRepo.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface NewsArticleRepo extends JpaRepository<NewsArticle, Long> {
+public interface NewsArticleRepo extends JpaRepository<NewsArticle, Integer> {
     Optional<NewsArticle> findNewsArticleByName (String name);
 }
